Separate fixtures module
